+++ juriscraper/opinions/united_states/state/delsuperct.py
@@ -14,7 +14,5 @@ class Site(delaware.Site):
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        self.url = (
-            "http://courts.delaware.gov/opinions/List.aspx?ag=Superior%20Court"
-        )
+        self.url = "https://courts.delaware.gov/opinions/List.aspx?ag=Superior%20Court"
         self.court_id = self.__module__