+++ CHANGES.md
@@ -18,10 +18,10 @@ Features:
 -
 
 Changes:
--
+- 
 
 Fixes:
--
+- handle different table structures in `uscgcoca` scraper #1526
 
 ## Current