+++ CHANGES.md
@@ -18,14 +18,26 @@ Features:
 -
 
 Changes:
+-
+
+Fixes:
+-
+
+## Current
+
+**2.6.94 - 2025-10-28**
+
+Changes:
 - update `guam` domain #1635
 
 Fixes:
 - handle different table structures in `uscgcoca` scraper #1526
 - update cleanup_content method for `ny` #1637
 - add expected content types for document handling in 'texapp' scraper #1636
+- update `delaware`, `delch`, `delsuperct` and `delctcompl` scrapers and examples #1641
 
-## Current
+
+## Past
 
 **2.6.93 - 2025-10-17**
 
@@ -37,9 +49,7 @@ Fixes:
 - fix `cal` rename key in OriginatingCourtInformation for accuracy #1625
 - fix `texbizct` improve error handling for name and citation extraction from titles #1629
 - add `ncbizct` to init #1631
-- update `delaware`, `delch`, `delsuperct` and `delctcompl` scrapers and examples #1641
 
-## Past
 
 **2.6.92 - 2025-10-01**